% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no2.R
\docType{data}
\name{no2}
\alias{no2}
\title{Dati giornalieri relativi all'NO2 per tutte le stazioni sul territorio}
\format{
Un tibble con dati da gennaio 2016 a giugno 2020 con 5 variabili:
\describe{
\item{reporting_year}{int anno della rilevazione}
\item{pollutant_fk}{int chiave esterna dell'inquinante}
\item{station_eu_code}{char codice europeo della stazione}
\item{date}{date data della rilevazione}
\item{value}{dbl dato osservato}
}
}
\usage{
no2
}
\description{
Per ottenere i dati regionali degli inquinani è sufficiente utilizzare le
funzionalità del pacchetto dplyr
}
\details{
Es: per i dati relativi al PM10 della regione Piemonte

library(dplyr)
filter(stazioniAria, region_id == 1) \%>\% inner_join(pm10, by = c("station_eu_code"))
}
\keyword{datasets}
