% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o3.R
\docType{data}
\name{o3}
\alias{o3}
\title{o3: massime orarie giornaliere e media mobile sulle
8 ore precedenti per tutte le stazioni sul territorio}
\format{
Un tibble con dati da gennaio 2016 a giugno 2020 con 5 variabili:
\describe{
\item{reporting_year}{Anno della rilevazione}
\item{pollutant_fk}{Chiave esterna dell'inquinante}
\item{station_eu_code}{Codice europeo della stazione}
\item{date}{Data della rilevazione}
\item{o3_max_h_d}{Massimo orario giornaliero}
\item{o3_max_mm8h_d}{Media mobile sulle 8 ore precedent}
}
}
\usage{
o3
}
\description{
Per ottenere i dati regionali degli inquinanti è sufficiente utilizzare le
funzionalità del pacchetto dplyr
}
\examples{
# i dati relativi al o3 della regione Piemonte 
library(dplyr)
dplyr::filter(stazioniAria, region_id == 1) \%>\% 
   inner_join(o3, by = c('station_eu_code'))

}
\keyword{datasets}
