% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm10.R
\docType{data}
\name{pm10}
\alias{pm10}
\title{Dati giornalieri relativi al PM10 per tutte le stazioni sul territorio}
\format{
Un tibble con dati da gennaio 2016 a giugno 2020 con 5 variabili:
\describe{
\item{reporting_year}{int anno della rilevazione}
\item{pollutant_fk}{int chiave esterna dell'inquinante}
\item{station_eu_code}{char codice europeo della stazione}
\item{date}{date data della rilevazione}
\item{value}{dbl dato osservato}
}
}
\usage{
pm10
}
\description{
Dati giornalieri relativi al PM10 per tutte le stazioni sul territorio
}
\keyword{datasets}
